rename the folder name to remove hyphen and move prepare_data.py one level up
